Formacoes/Apache_Kafka/kafka_producers_consumers_streams/ecommerce/pyproject.toml
[project]
name = "ecommerce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kafka-python>=2.2.6",
]
